Fixing randomization per Fervidusletum's comment
"This was a gotcha for me too. Declaring a new random each time you call the randomize method can lead to some undesired behavior. Since the default behavior of new random is to use time for the seed, quick repeated calls to this function can result in the same seed and getting the same answer ordering for a few calls in a row. One way to get around this is to only make one random and pass it as an arg to this method."